self-hosted-server/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
supabase==2.9.1
jcs==0.2.1
pydantic==2.9.2
python-dotenv==1.0.1
